[CI] Force to use universal GR binaries when building macOS wheels
cibuildwheel creates macOS x64 wheels by cross compiling on Apple
Silicon. This leads to linking problems with GR binaries since python-gr
cannot detect that it is used in a cross compile and would need the x64
GR binaries. To avoid this problem, force to use the universal build of
GR.